requirements.txt
aiofiles==0.8.0
alembic==1.8.0
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.1
fastapi-debug-toolbar==0.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
orjson==3.6.7
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyinstrument==4.1.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
sqlparse==0.4.2
sqltap==0.3.11
starlette==0.17.1
typing-extensions==4.1.1
ujson==4.3.0
urllib3==1.26.9
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
Werkzeug==2.1.2
